refactor: 로또 구매 및 생성 로직 별도의 컨트롤러로 분리
